LayoutTests/imported/w3c/web-platform-tests/intersection-observer/v2/z-index-changes-expected.txt



FAIL IntersectionObserver observes visibility changes from z-index assert_equals: Initial observation should be visible expected (boolean) true but got (undefined) undefined
